feat(FN-1865): move mock data back to utils
