--- e2e-tests/e2e-fixtures/get-cypress-env-variable.fixture.ts
@@ -1,4 +1,0 @@
-import { CypressEnvKey, CypressEnvVariable } from '../support/cypress-env';
-
-export const getCypressEnvVariable = <Key extends CypressEnvKey>(cypressEnvKey: Key): CypressEnvVariable<Key> =>
-  Cypress.env(cypressEnvKey) as CypressEnvVariable<Key>;
