+++ e2e-tests/gef/cypress.config.js
@@ -1,6 +1,5 @@
 const { defineConfig } = require('cypress');
 const { createTasks } = require('../support/tasks');
-const { CYPRESS_ENV } = require('../support/cypress-env');
 
 module.exports = defineConfig({
   dealApiProtocol: 'http://',
@@ -32,5 +31,4 @@ module.exports = defineConfig({
     },
   },
   experimentalCspAllowList: ['child-src', 'default-src', 'frame-src', 'form-action', 'script-src', 'script-src-elem'],
-  env: CYPRESS_ENV,
 });
