+++ libs/common/src/sql-db-seeder/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from './get-mongo-db-users';
+export * from './get-from-mongo-db';
