--- libs/common/src/test-helpers/mock-data/banks.mock.ts
@@ -1,165 +1,0 @@
-import { OptionalId } from 'mongodb';
-import { Bank } from '../../types/mongo-db-models';
-
-const MONTHLY_REPORT_PERIOD_SCHEDULE = [
-  { startMonth: 1, endMonth: 1 },
-  { startMonth: 2, endMonth: 2 },
-  { startMonth: 3, endMonth: 3 },
-  { startMonth: 4, endMonth: 4 },
-  { startMonth: 5, endMonth: 5 },
-  { startMonth: 6, endMonth: 6 },
-  { startMonth: 7, endMonth: 7 },
-  { startMonth: 8, endMonth: 8 },
-  { startMonth: 9, endMonth: 9 },
-  { startMonth: 10, endMonth: 10 },
-  { startMonth: 11, endMonth: 11 },
-  { startMonth: 12, endMonth: 12 },
-];
-
-export const MOCK_BANKS: OptionalId<Bank>[] = [
-  {
-    id: '956',
-    name: 'Barclays Bank',
-    mga: ['Test.pdf'],
-    emails: ['maker4@ukexportfinance.gov.uk', 'checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '01026167',
-    partyUrn: '00300130',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'Barclays Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: true,
-  },
-  {
-    id: '961',
-    name: 'HSBC',
-    mga: ['Test.pdf'],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '00014259',
-    partyUrn: '00300342',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'HSBC Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: true,
-  },
-  {
-    id: '999',
-    name: 'GEF Only Bank',
-    mga: ['Test.pdf'],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '00010452',
-    partyUrn: '00300312',
-    hasGefAccessOnly: true,
-    paymentOfficerTeam: {
-      teamName: 'GEF Only Bank Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: false,
-  },
-  {
-    id: '5213',
-    name: 'Newable',
-    mga: [''],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '08796862',
-    partyUrn: '00319394',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'Newable Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: false,
-  },
-  {
-    id: '964',
-    name: 'LLOYDS',
-    mga: ['Test.pdf'],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '00002065',
-    partyUrn: '00300302',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'LLOYDS Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: [
-      { startMonth: 3, endMonth: 5 },
-      { startMonth: 6, endMonth: 8 },
-      { startMonth: 9, endMonth: 11 },
-      { startMonth: 12, endMonth: 2 },
-    ],
-    isVisibleInTfmUtilisationReports: false,
-  },
-  {
-    id: '1004',
-    name: 'RBS',
-    mga: ['Mock.pdf'],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: 'SC090312',
-    partyUrn: '00307351',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'RBS Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: false,
-  },
-  {
-    id: '953',
-    name: 'Santander',
-    mga: ['Mock.pdf'],
-    emails: ['checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: '02294747',
-    partyUrn: '00311081',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'Santander Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: true,
-  },
-  {
-    id: '9',
-    name: 'UKEF test bank (Delegated)',
-    mga: ['mga_ukef_1.docx', 'mga_ukef_2.docx'],
-    emails: ['maker1@ukexportfinance.gov.uk', 'checker1@ukexportfinance.gov.uk'],
-    companiesHouseNo: 'UKEF0001',
-    partyUrn: '00318345',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'UKEF test bank (Delegated) Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: MONTHLY_REPORT_PERIOD_SCHEDULE,
-    isVisibleInTfmUtilisationReports: true,
-  },
-  {
-    id: '10',
-    name: 'UKEF test bank (Delegated) 2',
-    mga: ['mga_ukef_1.docx', 'mga_ukef_2.docx'],
-    emails: ['maker4@ukexportfinance.gov.uk', 'checker4@ukexportfinance.gov.uk'],
-    companiesHouseNo: 'UKEF0001',
-    partyUrn: '00318345',
-    hasGefAccessOnly: false,
-    paymentOfficerTeam: {
-      teamName: 'UKEF test bank (Delegated) 2 Payment Reporting Team',
-      email: 'payment-officer4@ukexportfinance.gov.uk',
-    },
-    utilisationReportPeriodSchedule: [
-      { startMonth: 3, endMonth: 5 },
-      { startMonth: 6, endMonth: 8 },
-      { startMonth: 9, endMonth: 11 },
-      { startMonth: 12, endMonth: 2 },
-    ],
-    isVisibleInTfmUtilisationReports: true,
-  },
-] as const;
