+++ libs/common/src/test-helpers/mock-data/index.ts
@@ -1,4 +1,2 @@
 export * from './utilisation-report.entity.mock';
-export * from './banks.mock';
 export * from './azure-file-info.mock';
-export * from './portal-users.mock';
